Fix PR #67 review findings: int64 session IDs, config consistency, flaky test
- model.go/db.go: revert charging-session IDs from int back to int64. The
  release builds armhf (GOARCH=arm, 32-bit) where int is 32 bits, so a large
  backend session ID would overflow on JSON unmarshal and on the bbolt key
  round-trip (Atoi/Itoa -> ParseInt/FormatInt). Add a large-ID regression test.
- config.go: use cs.Model() instead of cs.Storage.Model() in the token-refresh
  interval accessors, matching the other 40+ Service methods.
- easee_internal_test.go: de-flake the token-refresh test. Poll for repeated
  Ping calls via a race-safe atomic counter instead of asserting an exact,
  timing-dependent count after a fixed 400ms sleep.

+++ src/internal/model/model.go
@@ -559,13 +559,13 @@ type TimestampedValue[T any] struct {
 }
 
 type StartChargingSession struct {
-	ID         int       `json:"Id"`
+	ID         int64     `json:"Id"`
 	MeterValue float64   `json:"MeterValue"`
 	Start      time.Time `json:"Start"`
 }
 
 type StopChargingSession struct {
-	ID              int       `json:"Id"`
+	ID              int64     `json:"Id"`
 	Energy          float64   `json:"EnergyKwh"`
 	MeterValueStart float64   `json:"MeterValueStart"`
 	MeterValueStop  float64   `json:"MeterValueStop"`
